Adjusted the ways in the raw strings relating to some calculations were recreated.
Added a .is_real() method to Number instances to aid with complex calculations.

Added more solutions to class results:
 - Expression raised to power of Expression.
 - Expression raised to power of Variable.
 - Variable raised to power of Expression.

Added comparative to prevent Expression.value() from returning Expression type if result is Number.

Added a return value for e (2.71828...) to the the precision specified.

Added a a value for the sin() function.
 - Allows real numbers for now (solution to taylor series).